Share the coordinator e2e helpers and tighten the getState notes
Review follow-up, no behavior change.

`assertActivityRenders` and `assertParked` were copied verbatim into the new
persistence-host spec, and both specs reached into an activity iframe's input
with the same four-line chain. They now live in
`cypress/e2e/standalone/utils/coordinator.js` alongside `activityInput` and
`typeIntoActivity`, following the `e2e/tagSpecific/utils/` convention.

Also: record why a failed rebuild deliberately leaves the request consumed,
note that a null message id can never match (which is what routes later
answers to the error branch at all), correct two comments that called the
pre-rebuild blanking synchronous, add the ordering caveat a persistence
backend needs to `COORDINATION.md`, and fix the changeset's account of the
wasted core start.
