+++ .changeset/first-getstate-answer-wins.md
@@ -10,4 +10,4 @@ A page can hold more than one listener willing to answer that request. On a Rune
 
 A request now has a single answer: the first that carries state for this document wins, and later answers to it are ignored. Within a page load the coordinator's warehouse is the fresher of the two and answers first, so the reader keeps their work; on a boot the warehouse cannot answer — the first visit of a session, say — the persistence layer is the only answerer and restores as before. An answer with no state does not consume the request, so a host with nothing saved for an activity cannot shut out a better answer still on its way.
 
-Rebuilding the document twice also cost a second full core start on every restore, which is now spent only when a state actually arrives to rebuild from.
+Each answer also rebuilt the document from scratch, so on such a page every restore paid for a full core start whose result was immediately thrown away. Only the winning answer rebuilds now.
